Rename sync-producer to sync-manager
