--- cmd/sync/producer/Dockerfile
@@ -1,5 +1,0 @@
-FROM alpine
-RUN apk add --update --no-cache ca-certificates
-ADD cluster-registry-sync-producer /bin/sync-producer
-USER nobody
-ENTRYPOINT ["/bin/sync-producer"]
