+++ pkg/sync/manager/result.go
@@ -10,7 +10,7 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-package producer
+package manager
 
 import (
 	ctrl "sigs.k8s.io/controller-runtime"
